backend/pyproject.toml
[project]
name = "dreamcanvas-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "anthropic (>=0.76.0,<0.77.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "boto3 (>=1.42.30,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "openai (>=2.16.0,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.13,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

# ==================== Ruff Configuration ====================
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ==================== MyPy Configuration ====================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

# ==================== Pytest Configuration ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"